frontend/requirements.txt
chromadb
google-generativeai
keybert
langchain
langchain-chroma
langchain-community
langchain-core
networkx
pypdf
rank-bm25
requests
scikit-learn
sentence-transformers
streamlit
tenacity
tiktoken
tqdm